// src/app/core/models/user.ts
export interface User {
  address: string;
  bank_salaty: string;
  bio: string;
  birthday: string;
  email: string;
  first_name: string;
  gender: number;
  employee_id: number;
  join_date: string;
  last_name: string;
  left_date: string;
  personal_photo: string;
  phone: string;
  position: string;
  salary: number;
  social_accounts: string;
}